requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
python-dotenv
pytz
sqlparse
tzdata
